requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
motor>=3.6.0
pymongo>=4.10.0
PyJWT>=2.8.0
passlib>=1.7.4
bcrypt>=4.0.0
google-genai>=0.2.2
python-dotenv>=1.0.1
pydantic>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
dnspython>=2.3.0
